ADD: ObjectExtension+
- extension으로 처리한 오브젝트들의 파일명 수정
