--- dormitoryFamiles/dormitoryFamiles/Extension/ExtenstionColor.swift
@@ -1,14 +1,0 @@
-//
-//  ExtenstionColor.swift
-//  dormitoryFamiles
-//
-//  Created by leehwajin on 2023/09/05.
-//
-
-
-import UIKit
-
-extension UIColor {
-    static let titleColor: UIColor =  #colorLiteral(red: 0.8296361566, green: 0.1152015552, blue: 0.3670018315, alpha: 1)
-    static let kakaoColor: UIColor = .kakaoColor
-}
