+++ dormitoryFamiles/dormitoryFamiles/Extension/UILabel+.swift
@@ -14,4 +14,25 @@ extension UILabel {
         string.addAttribute(NSAttributedString.Key.kern, value: kernalValue, range: NSRange(location: 0, length: string.length-1))
         attributedText = string
     }
+    
+    func lineSpacing(_ spacing: CGFloat) {
+        guard let text = text, let font = font else { return }
+        let paragraphStyle = NSMutableParagraphStyle()
+        paragraphStyle.lineSpacing = spacing
+        let attributedString = NSAttributedString(string: text, attributes: [
+            .font: font,
+            .paragraphStyle: paragraphStyle
+        ])
+        self.attributedText = attributedString
+    }
+    
+    func asColor(targetString: [String], color: UIColor) {
+        let fullText = text ?? ""
+        let attributedString = NSMutableAttributedString(string: fullText)
+        for str in targetString {
+            let range = (fullText as NSString).range(of: str)
+            attributedString.addAttribute(.foregroundColor, value: color, range: range)
+        }
+        attributedText = attributedString
+    }
 }
